SmartString - initial implementation of a shared string class implementation.
The SmartString contains a pointer to the string_ptr class, used to store the string itself. The string_ptr also contains our reference counting code.

Each time the SmartString is copied the reference count is increased. The destruction or assignment of a SmartString results in the reference count being deceremented.

If the reference count reaches 0, the underlying string_ptr is deleted.

The class also keeps track of pointer references to the underlying string. If the string has potentially been modified we must make a deep copy to prevent other users of the original string from being updated. Also this potential update renders the string unshareable (as it could be modified at any time).

If the string can not be shared then the behaviour of the copy constructor and assignment operators must take this into account, i.e. an unshareable string can only ever have one reference.

There are 2 main classes:
SmartString - the main implementation class.
string_ptr - the structure is used as a pointer within SmartString and contains the reference count and the shareable state.