Renamed ReferenceCounted to RCObject. Renamed ReferenceCounted::IncrementReference to addReference Renamed ReferenceCounted::DecrementReference to removeReference.
Added a class 'shared_ptr' which is used to encapsulate the 'string_ptr' in the main String class.

The reference counting and management of the reference counted object is then managed by actions on the shared_ptr object, essentially removing all RCObject manipulations out of the the String class.

One exception remains which is the non const version of the index operator, String::operator[](size_t index);